% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DreamTheatR.R
\docType{package}
\name{DreamTheatR}
\alias{DreamTheatR}
\title{DreamTheatR}
\description{
list of palettes inspired by Dream Theater album artwork
}
\details{
list of palettes inspired by Dream Theater album artwork
}
