% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dt_palette}
\alias{dt_palette}
\title{A Dream Theater palette generator}
\usage{
dt_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{DramaticTurn1}, \code{Falling1},  \code{Images1},
\code{SixDegrees1}, \code{ViewFromTheTop1}}

\item{n}{Number of colors desired. If omitted, uses all colors.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
These are a handful of color palettes inspired by Dream Theater album artwork
}
\examples{
dt_palette("DramaticTurn1")
dt_palette("SixDegrees1")
dt_palette("ViewFromTheTop1")
dt_palette("ViewFromTheTop1", 3)

# If you need more colors than normally found in a palette, you
# can use a continuous palette to interpolate between existing colors
pal <- dt_palette(21, name = "DramaticTurn1", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
