% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{dt_palettes}
\alias{dt_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
dt_palettes
}
\description{
Use \code{\link{dt_palette}} to construct palettes of desired length.
}
\keyword{datasets}
